chore: 방명록 관련 로직 위치 이동
